DAOS-6901 test: Re-enable osa offline drain test Test-tag-hw-medium: pr,hw,medium,ib2 osa
