+++ src/tests/ftest/osa/osa_offline_drain.yaml
@@ -48,8 +48,10 @@ pool:
     svcn: 4
     control_method: dmg
 container:
-  properties:
-    enable_checksum: True
+    type: POSIX
+    control_method: daos
+    oclass: RP_2G1
+    properties: cksum:crc64,cksum_size:16384,srv_cksum:on
 dkeys:
   single:
     no_of_dkeys:
@@ -62,3 +64,21 @@ record:
   1KB:
     length:
       - 1024
+ior:
+    clientslots:
+      slots: 48
+    test_file: /testFile
+    repetitions: 1
+    dfs_destroy: False
+    iorflags:
+      write_flags: "-w -F -k -G 1"
+      read_flags: "-F -r -R -k -G 1"
+      ior_api:
+        - DFS
+      obj_class:
+        - RP_2G1
+    ior_test_sequence:
+    #   - [scmsize, nvmesize, transfersize, blocksize]
+    #    The values are set to be in the multiples of 10.
+    #    Values are appx GB.
+      - [6000000000, 54000000000, 500000, 500000000]
