+++ src/tests/ftest/osa/osa_offline_reintegration.py
@@ -42,37 +42,6 @@ class OSAOfflineReintegration(OSAUtils):
             self.hostlist_clients, self.workdir, None)
         self.out_queue = queue.Queue()
 
-    def run_ior_thread(self, action, oclass, api, test):
-        """Start the IOR thread for either writing or
-        reading data to/from a container.
-        Args:
-            action (str): Start the IOR thread with Read or
-                          Write
-            oclass (str): IOR object class
-            API (str): IOR API
-            test (list): IOR test sequence
-            flags (str): IOR flags
-        """
-        if action == "Write":
-            flags = self.ior_w_flags
-        else:
-            flags = self.ior_r_flags
-
-        # Add a thread for these IOR arguments
-        process = threading.Thread(target=self.ior_thread,
-                                   kwargs={"pool": self.pool,
-                                           "oclass": oclass,
-                                           "api": api,
-                                           "test": test,
-                                           "flags": flags,
-                                           "results":
-                                           self.out_queue})
-        # Launch the IOR thread
-        process.start()
-        # Wait for the thread to finish
-        process.join()
-
-
     def run_offline_reintegration_test(self, num_pool, data=False,
                                        server_boot=False):
         """Run the offline reintegration without data.
